Add CreateOrderTest

+++ src/test/java/tests/CreateUserTest.java
@@ -79,6 +79,10 @@ public class CreateUserTest extends BaseTest {
 
     @After
     public void afterClass() throws Exception {
-        createUser.deleteUser(accessToken);
+        if (accessToken != null) {
+            createUser.deleteUser(accessToken);
+        } else {
+            System.out.println("User is null !");
+        }
     }
 }
